+++ src/Game/Werewolf/Response.hs
@@ -165,8 +165,8 @@ playerQuitMessage player = publicMessage $ T.unwords [player ^. name, "the", pla
 
 currentTurnMessage :: Text -> Turn -> Message
 currentTurnMessage name NoOne   = gameIsOverMessage name
-currentTurnMessage name turn    = privateMessage [name] $ T.unwords [
-    "It's currently the", T.pack $ show turn, "turn."
+currentTurnMessage name turn    = privateMessage [name] $ T.concat [
+    "It's currently the ", T.pack $ show turn, "' turn."
     ]
 
 rolesInGameMessage :: Maybe [Text] -> [Role] -> Message