+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 *Revisions*
 
 * Added missing apostrophe to the new turn message. ([#63](https://github.com/hjwylde/werewolf/issues/63))
+* Changed the "Whom would you like to lynch?" text to be a public message displayed after the devoured message. ([#56](https://github.com/hjwylde/werewolf/issues/56))
 
 #### v0.3.3.2